+++ NEWS.md
@@ -1,6 +1,7 @@
 # quanteda.textstats 0.93
 
 * Minor changes to ensure compatibility with **quanteda** v3.
+* Changes to avoid breaking tests on older releases, caused by changes to the default for `stringsAsFactors` in `data.frame()`.
 
 # quanteda.textstats 0.92